backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google APIs
google-generativeai==0.3.1  # Gemini API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Security and authentication
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# CORS support
fastapi-cors==0.0.6
